move url to constant file
